refactor: extract ConversationGroupRepository from ConversationRepository
