+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/conversation/ConversationScope.kt
@@ -11,6 +11,7 @@ import com.wire.kalium.logic.data.team.TeamRepository
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.data.user.UserRepository
 import com.wire.kalium.logic.cache.SelfConversationIdProvider
+import com.wire.kalium.logic.data.conversation.ConversationGroupRepository
 import com.wire.kalium.logic.feature.connection.MarkConnectionRequestAsNotifiedUseCase
 import com.wire.kalium.logic.feature.connection.MarkConnectionRequestAsNotifiedUseCaseImpl
 import com.wire.kalium.logic.feature.connection.ObserveConnectionListUseCase
@@ -27,6 +28,7 @@ import com.wire.kalium.logic.sync.SyncManager
 @Suppress("LongParameterList")
 class ConversationScope internal constructor(
     private val conversationRepository: ConversationRepository,
+    private val conversationGroupRepository: ConversationGroupRepository,
     private val connectionRepository: ConnectionRepository,
     private val userRepository: UserRepository,
     private val syncManager: SyncManager,
@@ -75,13 +77,13 @@ class ConversationScope internal constructor(
         get() = DeleteTeamConversationUseCaseImpl(getSelfTeamUseCase, teamRepository, conversationRepository)
 
     val createGroupConversation: CreateGroupConversationUseCase
-        get() = CreateGroupConversationUseCase(conversationRepository, syncManager, clientRepository)
+        get() = CreateGroupConversationUseCase(conversationRepository, conversationGroupRepository, syncManager, clientRepository)
 
     val addMemberToConversationUseCase: AddMemberToConversationUseCase
-        get() = AddMemberToConversationUseCaseImpl(conversationRepository, selfUserId, persistMessage)
+        get() = AddMemberToConversationUseCaseImpl(conversationGroupRepository, selfUserId, persistMessage)
 
     val getOrCreateOneToOneConversationUseCase: GetOrCreateOneToOneConversationUseCase
-        get() = GetOrCreateOneToOneConversationUseCase(conversationRepository)
+        get() = GetOrCreateOneToOneConversationUseCase(conversationRepository, conversationGroupRepository)
 
     val updateConversationMutedStatus: UpdateConversationMutedStatusUseCase
         get() = UpdateConversationMutedStatusUseCaseImpl(conversationRepository)
@@ -108,10 +110,10 @@ class ConversationScope internal constructor(
         get() = UpdateConversationMemberRoleUseCaseImpl(conversationRepository)
 
     val removeMemberFromConversation: RemoveMemberFromConversationUseCase
-        get() = RemoveMemberFromConversationUseCaseImpl(conversationRepository, selfUserId, persistMessage)
+        get() = RemoveMemberFromConversationUseCaseImpl(conversationGroupRepository, selfUserId, persistMessage)
 
     val leaveConversation: LeaveConversationUseCase
-        get() = LeaveConversationUseCaseImpl(conversationRepository, selfUserId, persistMessage)
+        get() = LeaveConversationUseCaseImpl(conversationGroupRepository, selfUserId, persistMessage)
 
     val updateMLSGroupsKeyingMaterials: UpdateKeyingMaterialsUseCase
         get() = UpdateKeyingMaterialsUseCaseImpl(mlsConversationRepository, updateKeyingMaterialThresholdProvider)
