+++ logic/src/commonTest/kotlin/com/wire/kalium/logic/data/conversation/ConversationRepositoryTest.kt
@@ -1,7 +1,9 @@
 package com.wire.kalium.logic.data.conversation
 
 import app.cash.turbine.test
+import com.wire.kalium.cryptography.MLSClient
 import com.wire.kalium.logic.StorageFailure
+import com.wire.kalium.logic.data.client.MLSClientProvider
 import com.wire.kalium.logic.data.event.Event
 import com.wire.kalium.logic.data.id.ConversationId
 import com.wire.kalium.logic.data.id.GroupID
@@ -15,15 +17,12 @@ import com.wire.kalium.logic.framework.TestConversation
 import com.wire.kalium.logic.framework.TestUser
 import com.wire.kalium.logic.functional.Either
 import com.wire.kalium.logic.util.TimeParser
-import com.wire.kalium.logic.util.shouldFail
 import com.wire.kalium.logic.util.shouldSucceed
 import com.wire.kalium.network.api.base.authenticated.client.ClientApi
 import com.wire.kalium.network.api.base.authenticated.conversation.ConvProtocol
 import com.wire.kalium.network.api.base.authenticated.conversation.ConvProtocol.MLS
 import com.wire.kalium.network.api.base.authenticated.conversation.ConversationApi
-import com.wire.kalium.network.api.base.authenticated.conversation.ConversationMemberAddedDTO
 import com.wire.kalium.network.api.base.authenticated.conversation.ConversationMemberDTO
-import com.wire.kalium.network.api.base.authenticated.conversation.ConversationMemberRemovedDTO
 import com.wire.kalium.network.api.base.authenticated.conversation.ConversationMembersResponse
 import com.wire.kalium.network.api.base.authenticated.conversation.ConversationPagingResponse
 import com.wire.kalium.network.api.base.authenticated.conversation.ConversationResponse
@@ -34,8 +33,6 @@ import com.wire.kalium.network.api.base.authenticated.conversation.model.UpdateC
 import com.wire.kalium.network.api.base.authenticated.notification.EventContentDTO
 import com.wire.kalium.network.api.base.model.ConversationAccessDTO
 import com.wire.kalium.network.api.base.model.ConversationAccessRoleDTO
-import com.wire.kalium.network.api.base.model.ErrorResponse
-import com.wire.kalium.network.exceptions.KaliumException
 import com.wire.kalium.network.utils.NetworkResponse
 import com.wire.kalium.persistence.dao.ConversationDAO
 import com.wire.kalium.persistence.dao.ConversationEntity
@@ -52,12 +49,10 @@ import io.mockative.anything
 import io.mockative.classOf
 import io.mockative.configure
 import io.mockative.eq
-import io.mockative.fun2
 import io.mockative.given
 import io.mockative.matching
 import io.mockative.mock
 import io.mockative.once
-import io.mockative.thenDoNothing
 import io.mockative.verify
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.Channel
@@ -86,7 +81,10 @@ class ConversationRepositoryTest {
     private val userRepository = mock(UserRepository::class)
 
     @Mock
-    private val mlsConversationRepository = mock(classOf<MLSConversationRepository>())
+    private val mlsClient = mock(classOf<MLSClient>())
+
+    @Mock
+    private val mlsClientProvider = mock(classOf<MLSClientProvider>())
 
     @Mock
     private val conversationDAO = configure(mock(ConversationDAO::class)) { stubsUnitByDefault = true }
@@ -112,7 +110,7 @@ class ConversationRepositoryTest {
     fun setup() {
         conversationRepository = ConversationDataSource(
             userRepository,
-            mlsConversationRepository,
+            mlsClientProvider,
             conversationDAO,
             conversationApi,
             messageDAO,
@@ -171,16 +169,21 @@ class ConversationRepositoryTest {
             .whenInvoked()
             .thenReturn(flowOf(TestUser.SELF))
 
-        given(mlsConversationRepository)
-            .suspendFunction(mlsConversationRepository::hasEstablishedMLSGroup)
+        given(mlsClientProvider)
+            .suspendFunction(mlsClientProvider::getMLSClient)
             .whenInvokedWith(anything())
-            .thenReturn(Either.Right(true))
+            .thenReturn(Either.Right(mlsClient))
+
+        given(mlsClient)
+            .function(mlsClient::conversationExists)
+            .whenInvokedWith(anything())
+            .thenReturn(true)
 
         conversationRepository.insertConversationFromEvent(event)
 
-        verify(mlsConversationRepository)
-            .suspendFunction(mlsConversationRepository::hasEstablishedMLSGroup)
-            .with(eq(GROUP_ID))
+        verify(mlsClient)
+            .suspendFunction(mlsClient::conversationExists)
+            .with(eq(RAW_GROUP_ID))
             .wasInvoked(once)
 
         verify(conversationDAO)
@@ -416,142 +419,6 @@ class ConversationRepositoryTest {
     }
 
     @Test
-    fun givenSelfUserBelongsToATeam_whenCallingCreateGroupConversation_thenConversationIsCreatedAtBackendAndPersisted() = runTest {
-
-        given(conversationApi)
-            .suspendFunction(conversationApi::createNewConversation)
-            .whenInvokedWith(anything())
-            .thenReturn(NetworkResponse.Success(CONVERSATION_RESPONSE, emptyMap(), 201))
-
-        given(userRepository)
-            .coroutine { userRepository.observeSelfUser() }
-            .then { flowOf(TestUser.SELF) }
-
-        given(conversationDAO)
-            .suspendFunction(conversationDAO::insertConversation)
-            .whenInvokedWith(anything())
-            .thenDoNothing()
-
-        given(conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .whenInvokedWith(anything(), anything())
-            .thenDoNothing()
-
-        val result = conversationRepository.createGroupConversation(
-            GROUP_NAME,
-            listOf(TestUser.USER_ID),
-            ConversationOptions(protocol = ConversationOptions.Protocol.PROTEUS)
-        )
-
-        result.shouldSucceed { }
-
-        verify(conversationDAO)
-            .suspendFunction(conversationDAO::insertConversation)
-            .with(anything())
-            .wasInvoked(once)
-
-        verify(conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .with(anything(), anything())
-            .wasInvoked(once)
-    }
-
-    @Test
-    fun givenSelfUserDoesNotBelongToATeam_whenCallingCreateGroupConversation_thenConversationIsCreatedAtBackendAndPersisted() = runTest {
-
-        val selfUserWithoutTeam = TestUser.SELF.copy(teamId = null)
-
-        given(conversationApi)
-            .suspendFunction(conversationApi::createNewConversation)
-            .whenInvokedWith(anything())
-            .thenReturn(NetworkResponse.Success(CONVERSATION_RESPONSE, emptyMap(), 201))
-
-        given(userRepository)
-            .coroutine { userRepository.observeSelfUser() }
-            .then { flowOf(selfUserWithoutTeam) }
-
-        given(conversationDAO)
-            .suspendFunction(conversationDAO::insertConversation)
-            .whenInvokedWith(anything())
-            .thenDoNothing()
-
-        given(conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .whenInvokedWith(anything(), anything())
-            .thenDoNothing()
-
-        val result = conversationRepository.createGroupConversation(
-            GROUP_NAME,
-            listOf(TestUser.USER_ID),
-            ConversationOptions(protocol = ConversationOptions.Protocol.PROTEUS)
-        )
-
-        result.shouldSucceed { }
-
-        verify(conversationDAO)
-            .suspendFunction(conversationDAO::insertConversation)
-            .with(anything())
-            .wasInvoked(once)
-
-        verify(conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .with(anything(), anything())
-            .wasInvoked(once)
-    }
-
-    @Test
-    fun givenMLSProtocolIsUsed_whenCallingCreateGroupConversation_thenMLSGroupIsEstablished() = runTest {
-        val conversationResponse = CONVERSATION_RESPONSE.copy(protocol = MLS)
-
-        given(conversationApi)
-            .suspendFunction(conversationApi::createNewConversation)
-            .whenInvokedWith(anything())
-            .thenReturn(NetworkResponse.Success(conversationResponse, emptyMap(), 201))
-
-        given(userRepository)
-            .coroutine { userRepository.observeSelfUser() }
-            .then { flowOf(TestUser.SELF) }
-
-        given(conversationDAO)
-            .suspendFunction(conversationDAO::insertConversation)
-            .whenInvokedWith(anything())
-            .thenDoNothing()
-
-        given(conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .whenInvokedWith(any(), any())
-            .thenDoNothing()
-
-        given(mlsConversationRepository)
-            .suspendFunction(mlsConversationRepository::establishMLSGroup)
-            .whenInvokedWith(anything())
-            .then { Either.Right(Unit) }
-
-        val result = conversationRepository.createGroupConversation(
-            GROUP_NAME,
-            listOf(TestUser.USER_ID),
-            ConversationOptions(protocol = ConversationOptions.Protocol.MLS)
-        )
-
-        result.shouldSucceed { }
-
-        verify(conversationDAO)
-            .suspendFunction(conversationDAO::insertConversation)
-            .with(anything())
-            .wasInvoked(once)
-
-        verify(conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .with(anything(), anything())
-            .wasInvoked(once)
-
-        verify(mlsConversationRepository)
-            .suspendFunction(mlsConversationRepository::establishMLSGroup)
-            .with(anything())
-            .wasInvoked(once)
-    }
-
-    @Test
     fun givenUserHasKnownContactAndConversation_WhenGettingConversationDetailsByExistingConversation_ReturnTheCorrectConversation() =
         runTest {
             // given
@@ -678,185 +545,6 @@ class ConversationRepositoryTest {
 
         conversationRepository.fetchConversationIfUnknown(conversationId)
             .shouldSucceed()
-    }
-
-    @Test
-    fun givenAConversationAndAPISucceedsWithChange_whenAddingMembersToConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(PROTEUS_PROTOCOL_INFO)
-            .withFetchUsersIfUnknownByIdsSuccessful()
-            .withAddMemberAPISucceedChanged()
-            .withSuccessfulMemberInsert()
-            .arrange()
-
-        conversationRepository.addMembers(listOf(TestConversation.USER_1), TestConversation.ID)
-            .shouldSucceed()
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .with(anything(), anything())
-            .wasInvoked(exactly = once)
-    }
-
-    @Test
-    fun givenAConversationAndAPISucceedsWithoutChange_whenAddingMembersToConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(PROTEUS_PROTOCOL_INFO)
-            .withFetchUsersIfUnknownByIdsSuccessful()
-            .withAddMemberAPISucceedUnchanged()
-            .withSuccessfulMemberInsert()
-            .arrange()
-
-        conversationRepository.addMembers(listOf(TestConversation.USER_1), TestConversation.ID)
-            .shouldSucceed()
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .with(anything(), anything())
-            .wasInvoked(exactly = once)
-    }
-
-    @Test
-    fun givenAConversationAndAPIFailed_whenAddingMembersToConversation_thenShouldNotSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(PROTEUS_PROTOCOL_INFO)
-            .withAddMemberAPIFailed()
-            .arrange()
-
-        conversationRepository.addMembers(listOf(TestConversation.USER_1), TestConversation.ID)
-            .shouldFail()
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-            .with(any(), any())
-            .wasNotInvoked()
-    }
-
-    @Test
-    fun givenAnMLSConversationAndAPISucceeds_whenAddMemberFromConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(MLS_PROTOCOL_INFO)
-            .withAddMemberAPISucceedChanged()
-            .withSuccessfulMemberInsert()
-            .withSuccessfulAddMemberToMLSGroup()
-            .arrange()
-
-        conversationRepository.addMembers(listOf(TestConversation.USER_1), TestConversation.ID)
-            .shouldSucceed { it is MemberChangeResult.Changed }
-
-        // this function called in the mlsRepo
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::deleteMemberByQualifiedID)
-            .with(anything(), anything())
-            .wasNotInvoked()
-        verify(arrangement.mlsConversationRepository)
-            .suspendFunction(arrangement.mlsConversationRepository::addMemberToMLSGroup)
-            .with(eq(GROUP_ID), eq(listOf(TestConversation.USER_1)))
-            .wasInvoked(exactly = once)
-    }
-
-    @Test
-    fun givenAConversationAndAPISucceedsWithChange_whenRemovingMemberFromConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(PROTEUS_PROTOCOL_INFO)
-            .withDeleteMemberAPISucceedChanged()
-            .withSuccessfulMemberDeletion()
-            .arrange()
-
-        conversationRepository.deleteMember(TestConversation.USER_1, TestConversation.ID)
-            .shouldSucceed { it is MemberChangeResult.Changed }
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::deleteMemberByQualifiedID)
-            .with(anything(), anything())
-            .wasInvoked(exactly = once)
-    }
-
-    @Test
-    fun givenAConversationAndAPISucceedsWithoutChange_whenRemovingMemberFromConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(PROTEUS_PROTOCOL_INFO)
-            .withDeleteMemberAPISucceedUnchanged()
-            .withSuccessfulMemberDeletion()
-            .arrange()
-
-        conversationRepository.deleteMember(TestConversation.USER_1, TestConversation.ID)
-            .shouldSucceed { it is MemberChangeResult.Unchanged }
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::deleteMemberByQualifiedID)
-            .with(anything(), anything())
-            .wasInvoked(exactly = once)
-    }
-
-    @Test
-    fun givenAConversationAndAPIFailed_whenRemovingMemberFromConversation_thenShouldFail() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(PROTEUS_PROTOCOL_INFO)
-            .withDeleteMemberAPIFailed()
-            .withSuccessfulMemberDeletion()
-            .arrange()
-
-        conversationRepository.deleteMember(TestConversation.USER_1, TestConversation.ID)
-            .shouldFail()
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::deleteMemberByQualifiedID)
-            .with(anything(), anything())
-            .wasNotInvoked()
-    }
-
-    @Test
-    fun givenAnMLSConversationAndAPISucceeds_whenRemovingLeavingConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(MLS_PROTOCOL_INFO)
-            .withDeleteMemberAPISucceedChanged()
-            .withSuccessfulMemberDeletion()
-            .withSuccessfulLeaveMLSGroup()
-            .arrange()
-
-        conversationRepository.deleteMember(TestUser.SELF.id, TestConversation.ID)
-            .shouldSucceed { it is MemberChangeResult.Changed }
-
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::deleteMemberByQualifiedID)
-            .with(anything(), anything())
-            .wasInvoked(exactly = once)
-        verify(arrangement.mlsConversationRepository)
-            .suspendFunction(arrangement.mlsConversationRepository::leaveGroup)
-            .with(eq(GROUP_ID))
-            .wasInvoked(exactly = once)
-        verify(arrangement.mlsConversationRepository)
-            .suspendFunction(arrangement.mlsConversationRepository::removeMembersFromMLSGroup)
-            .with(any(), any())
-            .wasNotInvoked()
-    }
-
-    @Test
-    fun givenAnMLSConversationAndAPISucceeds_whenRemoveMemberFromConversation_thenShouldSucceed() = runTest {
-        val (arrangement, conversationRepository) = Arrangement()
-            .withConversationProtocolIs(MLS_PROTOCOL_INFO)
-            .withDeleteMemberAPISucceedChanged()
-            .withSuccessfulMemberDeletion()
-            .withSuccessfulRemoveMemberFromMLSGroup()
-            .arrange()
-
-        conversationRepository.deleteMember(TestConversation.USER_1, TestConversation.ID)
-            .shouldSucceed { it is MemberChangeResult.Changed }
-
-        // this function called in the mlsRepo
-        verify(arrangement.conversationDAO)
-            .suspendFunction(arrangement.conversationDAO::deleteMemberByQualifiedID)
-            .with(anything(), anything())
-            .wasNotInvoked()
-        verify(arrangement.mlsConversationRepository)
-            .suspendFunction(arrangement.mlsConversationRepository::removeMembersFromMLSGroup)
-            .with(eq(GROUP_ID), eq(listOf(TestConversation.USER_1)))
-            .wasInvoked(exactly = once)
-        verify(arrangement.mlsConversationRepository)
-            .suspendFunction(arrangement.mlsConversationRepository::leaveGroup)
-            .with(any())
-            .wasNotInvoked()
     }
 
     @Suppress("LongMethod")
@@ -1362,7 +1050,10 @@ class ConversationRepositoryTest {
         val userRepository: UserRepository = mock(UserRepository::class)
 
         @Mock
-        val mlsConversationRepository: MLSConversationRepository = mock(MLSConversationRepository::class)
+        val mlsClient: MLSClient = mock(MLSClient::class)
+
+        @Mock
+        val mlsClientProvider: MLSClientProvider = mock(MLSClientProvider::class)
 
         @Mock
         val conversationDAO: ConversationDAO = mock(ConversationDAO::class)
@@ -1385,7 +1076,7 @@ class ConversationRepositoryTest {
         val conversationRepository =
             ConversationDataSource(
                 userRepository,
-                mlsConversationRepository,
+                mlsClientProvider,
                 conversationDAO,
                 conversationApi,
                 messageDAO,
@@ -1441,131 +1132,6 @@ class ConversationRepositoryTest {
                 .thenReturn(Unit)
         }
 
-        fun withConversationProtocolIs(protocolInfo: ConversationEntity.ProtocolInfo) = apply {
-            given(conversationDAO)
-                .suspendFunction(conversationDAO::getConversationByQualifiedID)
-                .whenInvokedWith(any())
-                .thenReturn(TestConversation.GROUP_ENTITY(protocolInfo))
-        }
-
-        fun withAddMemberAPISucceedChanged() = apply {
-            given(conversationApi)
-                .suspendFunction(conversationApi::addMember)
-                .whenInvokedWith(any(), any())
-                .thenReturn(
-                    NetworkResponse.Success(
-                        TestConversation.ADD_MEMBER_TO_CONVERSATION_SUCCESSFUL_RESPONSE,
-                        mapOf(),
-                        HttpStatusCode.OK.value
-                    )
-                )
-        }
-
-        fun withAddMemberAPISucceedUnchanged() = apply {
-            given(conversationApi)
-                .suspendFunction(conversationApi::addMember)
-                .whenInvokedWith(any(), any())
-                .thenReturn(
-                    NetworkResponse.Success(
-                        ConversationMemberAddedDTO.Unchanged,
-                        mapOf(),
-                        HttpStatusCode.OK.value
-                    )
-                )
-        }
-
-        fun withAddMemberAPIFailed() = apply {
-            given(conversationApi)
-                .suspendFunction(conversationApi::addMember)
-                .whenInvokedWith(any(), any())
-                .thenReturn(
-                    NetworkResponse.Error(
-                        KaliumException.ServerError(ErrorResponse(500, "error_message", "error_label"))
-                    )
-                )
-
-        }
-
-        fun withDeleteMemberAPISucceedChanged() = apply {
-            given(conversationApi)
-                .suspendFunction(conversationApi::removeMember)
-                .whenInvokedWith(any(), any())
-                .thenReturn(
-                    NetworkResponse.Success(
-                        TestConversation.REMOVE_MEMBER_FROM_CONVERSATION_SUCCESSFUL_RESPONSE,
-                        mapOf(),
-                        HttpStatusCode.OK.value
-                    )
-                )
-        }
-
-        fun withDeleteMemberAPISucceedUnchanged() = apply {
-            given(conversationApi)
-                .suspendFunction(conversationApi::removeMember)
-                .whenInvokedWith(any(), any())
-                .thenReturn(
-                    NetworkResponse.Success(
-                        ConversationMemberRemovedDTO.Unchanged,
-                        mapOf(),
-                        HttpStatusCode.OK.value
-                    )
-                )
-        }
-
-        fun withDeleteMemberAPIFailed() = apply {
-            given(conversationApi)
-                .suspendFunction(conversationApi::removeMember)
-                .whenInvokedWith(any(), any())
-                .thenReturn(
-                    NetworkResponse.Error(
-                        KaliumException.ServerError(ErrorResponse(500, "error_message", "error_label"))
-                    )
-                )
-
-        }
-
-        fun withFetchUsersIfUnknownByIdsSuccessful() = apply {
-            given(userRepository)
-                .suspendFunction(userRepository::fetchUsersIfUnknownByIds)
-                .whenInvokedWith(any())
-                .thenReturn(Either.Right(Unit))
-        }
-
-        fun withSuccessfulMemberInsert() = apply {
-            given(conversationDAO)
-                .suspendFunction(conversationDAO::insertMembersWithQualifiedId, fun2<List<MemberEntity>, QualifiedIDEntity>())
-                .whenInvokedWith(any(), any())
-                .thenDoNothing()
-        }
-
-        fun withSuccessfulMemberDeletion() = apply {
-            given(conversationDAO)
-                .suspendFunction(conversationDAO::deleteMemberByQualifiedID)
-                .whenInvokedWith(any(), any())
-                .thenReturn(Unit)
-        }
-
-        fun withSuccessfulLeaveMLSGroup() = apply {
-            given(mlsConversationRepository)
-                .suspendFunction(mlsConversationRepository::leaveGroup)
-                .whenInvokedWith(any())
-                .thenReturn(Either.Right(Unit))
-        }
-
-        fun withSuccessfulAddMemberToMLSGroup() = apply {
-            given(mlsConversationRepository)
-                .suspendFunction(mlsConversationRepository::addMemberToMLSGroup)
-                .whenInvokedWith(any(), any())
-                .thenReturn(Either.Right(Unit))
-        }
-
-        fun withSuccessfulRemoveMemberFromMLSGroup() = apply {
-            given(mlsConversationRepository)
-                .suspendFunction(mlsConversationRepository::removeMembersFromMLSGroup)
-                .whenInvokedWith(any(), any())
-                .thenReturn(Either.Right(Unit))
-        }
-
         fun withSuccessfulConversationDeletion() = apply {
             given(conversationDAO)
                 .suspendFunction(conversationDAO::deleteConversationByQualifiedID)
@@ -1610,16 +1176,6 @@ class ConversationRepositoryTest {
     companion object {
         private const val RAW_GROUP_ID = "mlsGroupId"
         val GROUP_ID = GroupID(RAW_GROUP_ID)
-        val PROTEUS_PROTOCOL_INFO = ConversationEntity.ProtocolInfo.Proteus
-        val MLS_PROTOCOL_INFO = ConversationEntity.ProtocolInfo
-            .MLS(
-                RAW_GROUP_ID,
-                groupState = ConversationEntity.GroupState.ESTABLISHED,
-                0UL,
-                Instant.parse("2021-03-30T15:36:00.000Z"),
-                cipherSuite = ConversationEntity.CipherSuite.MLS_128_DHKEMX25519_AES128GCM_SHA256_Ed25519
-            )
-
         const val GROUP_NAME = "Group Name"
 
         val CONVERSATION_ID = TestConversation.ID
